Adding Content and Styling
In this commit I added my first Draft to this challenge. All it includes is the content provided, and some styles to make sure my style sheet is properly linked. Additionally, I do not yet know how to make websites responsive so I've added the bootstrap framework to do the work for me.